CARL ResourceManager: Split event listener implementation into own module.
