+++ opendut-carl/src/resource/manager/mod.rs
@@ -1,4 +1,5 @@
 pub(super) mod persistence;
+mod event_listener;
 mod resources;
 
 pub use resources::Resources;
@@ -17,7 +18,6 @@ use std::sync::Arc;
 use config::Config;
 use tokio::sync::{RwLock, RwLockWriteGuard};
 use tokio_util::sync::CancellationToken;
-use tracing::{debug, warn};
 
 
 pub type ResourceManagerRef = Arc<ResourceManager>;
@@ -124,112 +124,22 @@ impl ResourceManager {
         &self,
         on_event: impl AsyncFn(SubscriptionEvent<R>) + Send + 'static,
     ) {
-        let resource_name = R::type_name();
-        let mut subscription = self.subscribe::<R>().await;
-        let cancel = self.event_listener_cancel.clone();
-
-        tokio::task::spawn_blocking(move || {
-            tokio::runtime::Handle::current().block_on(async move {
-                loop {
-                    tokio::select! {
-                        received = subscription.receive() => match received {
-                            Ok(event) => {
-                                on_event(event).await;
-                            }
-                            Err(subscription::ReceiveError::Broadcast(error)) => match error {
-                                tokio::sync::broadcast::error::RecvError::Closed => {
-                                    warn!("ResourceManager subscription channel for {resource_name} closed. Aborting.");
-                                    break;
-                                }
-                                tokio::sync::broadcast::error::RecvError::Lagged(skipped_messages) => {
-                                    warn!("ResourceManager subscription channel for {resource_name} lagged behind and had to skip {skipped_messages} messages.");
-                                },
-                            }
-                        },
-                        _ = cancel.cancelled() => {
-                            debug!("ResourceManager is cancelling subscription channel for {resource_name}.");
-                            break;
-                        }
-                    }
-                }
-            })
-        });
+        event_listener::spawn_event_listener(self, on_event).await
     }
 
 
-    //TODO move implementation into submodule
+    /// Run a closure when two resources are inserted/updated or removed.
+    ///
+    /// It aggregates the events of these resources, meaning you will
+    /// receive a tuple with the most recent event for each.
+    ///
+    /// The events are wrapped into Options, because when the first event
+    /// fires, there is no previous event for the rest.
     pub async fn spawn_event_listener_aggregate_2<R1: Resource + Subscribable, R2: Resource + Subscribable>(
         &self,
         on_event: impl AsyncFn((Option<SubscriptionEvent<R1>>, Option<SubscriptionEvent<R2>>)) + Send + 'static,
     ) {
-
-        let mut subscription1 = self.subscribe::<R1>().await;
-        let mut subscription2 = self.subscribe::<R2>().await;
-
-        let aggregate_resource_name = format!("({}, {})", R1::type_name(), R2::type_name());
-
-
-        let cancel = self.event_listener_cancel.clone();
-
-
-        enum Abort { Yes, No }
-
-        async fn receive<R: Resource>(subscription: &mut Subscription<R>) -> Result<SubscriptionEvent<R>, Abort> { //TODO maybe pull out and re-use for non-aggregate as well
-            let resource_name = R::type_name();
-
-            match subscription.receive().await {
-                Ok(event) => {
-                    Ok(event) //TODO simplify
-                }
-                Err(subscription::ReceiveError::Broadcast(error)) => match error {
-                    tokio::sync::broadcast::error::RecvError::Closed => {
-                        debug!("ResourceManager subscription channel for {resource_name} closed. Aborting."); //TODO mention aggregate channel name
-                        Err(Abort::Yes)
-                    }
-                    tokio::sync::broadcast::error::RecvError::Lagged(skipped_messages) => {
-                        warn!("ResourceManager subscription channel for {resource_name} lagged behind and had to skip {skipped_messages} messages.");
-                        Err(Abort::No)
-                    },
-                }
-            }
-        }
-
-
-        tokio::task::spawn_blocking(move || {
-            tokio::runtime::Handle::current().block_on(async move {
-                let mut previous1 = None;
-                let mut previous2 = None;
-
-                loop {
-                    tokio::select! {
-                        result = receive(&mut subscription1) => match result {
-                            Ok(event) => {
-                                previous1.replace(event);
-
-                                let notification = (previous1.clone(), previous2.clone());
-                                on_event(notification).await;
-                            }
-                            Err(Abort::Yes) => break,
-                            Err(Abort::No) => {}, //do nothing
-                        },
-                        result = receive(&mut subscription2) => match result {
-                            Ok(event) => {
-                                previous2.replace(event);  //TODO pull into closure?
-
-                                let notification = (previous1.clone(), previous2.clone());
-                                on_event(notification).await;
-                            }
-                            Err(Abort::Yes) => break,
-                            Err(Abort::No) => {}, //do nothing
-                        },
-                        _ = cancel.cancelled() => {
-                            debug!("ResourceManager is cancelling subscription channel for {aggregate_resource_name}.");
-                            break;
-                        }
-                    }
-                }
-            })
-        });
+        event_listener::spawn_event_listener_aggregate_2(self, on_event).await
     }
 
     /// Get events when a resource is inserted/updated or removed.
@@ -351,15 +261,12 @@ impl ResourceManager {
 #[cfg(test)]
 mod test {
     use std::collections::HashSet;
-    use std::time::Duration;
     use std::vec;
 
     use googletest::prelude::*;
-    use opendut_model::peer::state::PeerConnectionState;
-    use tokio::sync::mpsc;
 
     use super::*;
-    use opendut_model::cluster::{ClusterDeployment, ClusterDescriptor, ClusterId, ClusterName};
+    use opendut_model::cluster::{ClusterDescriptor, ClusterId, ClusterName};
     use opendut_model::peer::executor::{container::{ContainerCommand, ContainerImage, ContainerName, Engine}, ExecutorDescriptor, ExecutorDescriptors, ExecutorId, ExecutorKind};
     use opendut_model::peer::{PeerDescriptor, PeerId, PeerLocation, PeerName, PeerNetworkDescriptor};
     use opendut_model::topology::Topology;
@@ -441,77 +348,6 @@ mod test {
             PersistenceResult::Ok(())
         }).await??;
 
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn should_spawn_an_event_listener() -> Result<()> {
-        let (testee, _cancel) = ResourceManager::new_in_memory();
-
-        let (sender, mut receiver) = mpsc::channel(1);
-
-        testee.spawn_event_listener::<ClusterDeployment>(async move |_event| {
-            sender.send(()).await.unwrap();
-        }).await;
-
-        let id = ClusterId::random();
-        testee.insert(id, ClusterDeployment { id }).await?;
-
-        let received = tokio::time::timeout(
-            Duration::from_secs(5),
-            receiver.recv()
-        ).await?;
-
-        assert!(received.is_some());
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn should_spawn_an_event_listener_that_aggregates_two_resources() -> Result<()> {
-        let (testee, _cancel) = ResourceManager::new_in_memory();
-
-        let (sender, mut receiver) = mpsc::channel(1);
-
-        testee.spawn_event_listener_aggregate_2::<ClusterDeployment, PeerConnectionState>(
-            async move |event| {
-                sender.send(event).await.unwrap();
-            }
-        ).await;
-
-
-
-        let id1 = ClusterId::random();
-        let resource1 = ClusterDeployment { id: id1 };
-        testee.insert(id1, resource1.clone()).await?;
-
-        {
-            let received = tokio::time::timeout(
-                Duration::from_secs(5),
-                receiver.recv()
-            ).await?;
-
-            let (received1, received2) = received.unwrap();
-
-            assert_eq!(received1, Some(SubscriptionEvent::Inserted { id: id1, value: resource1.clone() }));
-            assert_eq!(received2, None);
-        }
-
-
-        let id2 = PeerId::random();
-        let resource2 = PeerConnectionState::Offline;
-        testee.insert(id2, resource2.clone()).await?;
-
-        {
-            let received = tokio::time::timeout(
-                Duration::from_secs(5),
-                receiver.recv()
-            ).await?;
-
-            let (received1, received2) = received.unwrap();
-
-            assert_eq!(received1, Some(SubscriptionEvent::Inserted { id: id1, value: resource1 }));
-            assert_eq!(received2, Some(SubscriptionEvent::Inserted { id: id2, value: resource2 }));
-        }
         Ok(())
     }
 }
